\name{d.biodiversity}

\docType{data}

\alias{d.biodiversity}

\title{Linkages and Collective Action in Biodiversity Conservation}

\description{
This dataset is from \href{http://dx.doi.org/10.1016/j.gloenvcha.2013.02.011}{Basurto (2013)}, who analyzes the determinants of the emergence and endurance of autonomy among local institutions for biodiversity conservation in Costa Rica using fsQCA. 
}

\usage{data(d.biodiversity)}

\format{
This data frame contains 30 rows (cases) and the following 9 columns (factors):

\tabular{rlll}{
[ , 1] \tab \bold{AU}  \tab endogenous factor:   \tab local autonomy                                                 \cr
       \tab            \tab            \tab ("1" always, "0" never)                                        \cr
[ , 2] \tab \bold{EM}  \tab exogenous factor: \tab local communal involvement through direct employment           \cr   
       \tab            \tab            \tab ("1" 100 percent, "0" 0 percent)                               \cr
[ , 3] \tab \bold{SP}  \tab exogenous factor: \tab local direct spending                                          \cr   
       \tab            \tab            \tab ("1" always, "0" never)                                        \cr
[ , 4] \tab \bold{CO}  \tab exogenous factor: \tab co-management with local or regional stakeholders              \cr
       \tab            \tab            \tab ("1" present, "0" absent)                                      \cr
[ , 5] \tab \bold{CI}  \tab exogenous factor: \tab degree of influence of national civil service policies         \cr
       \tab            \tab            \tab ("1" 100 percent civil service employees, "0" 0 percent)       \cr
[ , 6] \tab \bold{PO}  \tab exogenous factor: \tab national participation in policy-making                        \cr
       \tab            \tab            \tab ("1" perceived programme influence, "0" no perceived influence)\cr
[ , 7] \tab \bold{RE}  \tab exogenous factor: \tab research-oriented partnerships                                 \cr
       \tab            \tab            \tab ("1" many, "0" few)                                            \cr
[ , 8] \tab \bold{CN}  \tab exogenous factor: \tab conservation-oriented partnerships                             \cr
       \tab            \tab            \tab ("1" many, "0" few)                                            \cr
[ , 9] \tab \bold{DE}  \tab exogenous factor: \tab direct support by development organizations                    \cr
       \tab            \tab            \tab ("1" always, "0" never)                                                         
}
}

\author{
Alrik Thiem (\href{http://www.alrik-thiem.net}{Personal Website}; \href{https://www.researchgate.net/profile/Alrik_Thiem}{ResearchGate Website})
}

\section{Contributors}{\tabular{l}{
Thiem, Alrik: collection, documentation
}}

\source{
Basurto, Xavier. 2013. \dQuote{Linking Multi-Level Governance to Local Common-Pool 
Resource Theory using Fuzzy-Set Qualitative Comparative Analysis: Insights from 
Twenty Years of Biodiversity Conservation in Costa Rica.} \emph{Global Environmental Change} \strong{23} (3):573-87. DOI: \href{http://dx.doi.org/10.1016/j.gloenvcha.2013.02.011}{10.1016/j.gloenvcha.2013.02.011}.
}

\keyword{datasets}
