\name{d.education}

\docType{data}

\alias{d.education}

\title{Impact of New Public Management Instruments on PhD Education}

\description{
This dataset is from \href{http://dx.doi.org/10.1007/s10734-009-9264-3}{Schneider and Sadowski (2010)}, who analyze the determinants of PhD placement success for 14 economics departments using csQCA. 
}

\usage{data(d.education)}

\format{
This data frame contains 14 rows (cases) and the following 7 columns (factors):

\tabular{rllrll}{
[ , 1] \tab \bold{NPM1} \tab exogenous factor\tab: \tab local competition      \tab ("1" used, "0" not used)\cr
[ , 2] \tab \bold{NPM2} \tab exogenous factor\tab: \tab national competition   \tab ("1" used, "0" not used)\cr
[ , 3] \tab \bold{NPM3} \tab exogenous factor\tab: \tab transparency           \tab ("1" used, "0" not used)\cr
[ , 4] \tab \bold{NPM4} \tab exogenous factor\tab: \tab university regulations \tab ("1" used, "0" not used)\cr
[ , 5] \tab \bold{NPM5} \tab exogenous factor\tab: \tab target agreements      \tab ("1" used, "0" not used)\cr
[ , 6] \tab \bold{NPM6} \tab exogenous factor\tab: \tab state regulations      \tab ("1" used, "0" not used)\cr
[ , 7] \tab \bold{O}    \tab endogenous factor\tab:   \tab placement success      \tab ("1" yes, "0" no)
}
}

\author{
Alrik Thiem (\href{http://www.alrik-thiem.net}{Personal Website}; \href{https://www.researchgate.net/profile/Alrik_Thiem}{ResearchGate Website})
}

\section{Contributors}{\tabular{l}{
Thiem, Alrik: collection, documentation
}}

\source{
Schneider, Peter, and Dieter Sadowski. 2010. \dQuote{The Impact of New Public 
Management Instruments on PhD Education.} \emph{Higher Education} \strong{59} (5):543-65.
DOI: \href{http://dx.doi.org/10.1007/s10734-009-9264-3}{10.1007/s10734-009-9264-3}.
}

\keyword{datasets}
