\name{d.homeless}

\docType{data}

\alias{d.homeless}

\title{Outcomes of Homeless Social Movement Organizations}

\description{
This dataset is from \href{http://www.jstor.org/stable/3003888}{Cress and Snow (2000)}, who analyze the determinants of 
the outcomes attained by homeless social movement organizations using csQCA. 
}

\usage{data(d.homeless)}

\format{
This data frame contains 15 rows (cases) and the following 10 columns (factors):

\tabular{rllrll}{
[ , 1]  \tab \bold{VI}  \tab exogenous factor\tab: \tab viability          \tab ("1" present, "0" absent)\cr
[ , 2]  \tab \bold{DT}  \tab exogenous factor\tab: \tab disruptive tactics \tab ("1" present, "0" absent)\cr
[ , 3]  \tab \bold{SA}  \tab exogenous factor\tab: \tab sympathetic allies \tab ("1" present, "0" absent)\cr
[ , 4]  \tab \bold{CS}  \tab exogenous factor\tab: \tab city support       \tab ("1" present, "0" absent)\cr
[ , 5]  \tab \bold{DF}  \tab exogenous factor\tab: \tab diagnostic frame   \tab ("1" present, "0" absent)\cr
[ , 6]  \tab \bold{PF}  \tab exogenous factor\tab: \tab prognostic frame   \tab ("1" present, "0" absent)\cr
[ , 7]  \tab \bold{REP} \tab endogenous factor\tab:   \tab representation     \tab ("1" present, "0" absent)\cr
[ , 8]  \tab \bold{RES} \tab endogenous factor\tab:   \tab resources          \tab ("1" present, "0" absent)\cr
[ , 9]  \tab \bold{RIG} \tab endogenous factor\tab:   \tab rights             \tab ("1" present, "0" absent)\cr
[ , 10] \tab \bold{REL} \tab endogenous factor\tab:   \tab relief             \tab ("1" present, "0" absent)
}
}

\author{
Alrik Thiem (\href{http://www.alrik-thiem.net}{Personal Website}; \href{https://www.researchgate.net/profile/Alrik_Thiem}{ResearchGate Website})
}

\section{Contributors}{\tabular{l}{
Thiem, Alrik: collection, documentation
}}

\source{
Cress, Daniel M., and David A. Snow. 2000. \dQuote{The Outcomes of Homeless Mobilization: 
The Influence of Organization, Disruption, Political Mediation, and Framing.} 
\emph{American Journal of Sociology} 105 (4):1063-104. \href{http://www.jstor.org/stable/3003888}{Link}.
}

\keyword{datasets}
