\name{d.socialsecurity}

\docType{data}

\alias{d.socialsecurity}

\title{The Emergence of the Social Security State}

\description{
This dataset is from \href{http://www.jstor.org/stable/2096417}{Hicks, Misra and Ng (1995)}, who analyze the emergence of
social security programs in 15 industrializing countries during the period
1880-1930 using csQCA.
}

\usage{data(d.socialsecurity)}

\format{
This data frame contains 30 rows (cases) and the following 6 columns (factors):

\tabular{rllrll}{
[ , 1]  \tab \bold{LG}  \tab exogenous factor\tab: \tab liberal government         \tab ("1" present, "0" absent)\cr
[ , 2]  \tab \bold{CG}  \tab exogenous factor\tab: \tab Catholic government        \tab ("1" present, "0" absent)\cr
[ , 3]  \tab \bold{PS}  \tab exogenous factor\tab: \tab patriarchal state          \tab ("1" present, "0" absent)\cr
[ , 4]  \tab \bold{UD}  \tab exogenous factor\tab: \tab unitary democracy          \tab ("1" present, "0" absent)\cr
[ , 5]  \tab \bold{WM}  \tab exogenous factor\tab: \tab working-class mobilization \tab ("1" present, "0" absent)\cr
[ , 6]  \tab \bold{CO}  \tab endogenous factor\tab:   \tab consolidation              \tab ("1" present, "0" absent)
}
}

\author{
Alrik Thiem (\href{http://www.alrik-thiem.net}{Personal Website}; \href{https://www.researchgate.net/profile/Alrik_Thiem}{ResearchGate Website})
}

\section{Contributors}{\tabular{l}{
Thiem, Alrik: collection, documentation
}}

\source{
Hicks, Alexander, Joya Misra, and Tang N. Ng. 1995. \dQuote{The Programmatic 
Emergence of the Social Security State.} 
\emph{American Sociological Review} \strong{60} (3):329-49. \href{http://www.jstor.org/stable/2096417}{Link}.
}

\keyword{datasets}
