\name{QCApro-internal}

\alias{.onAttach}
\alias{base3rows}
\alias{createChart}
\alias{createString}
\alias{demoChart}
\alias{DeMorganLoop}
\alias{eQMCLoop}
\alias{findSubsets}
\alias{findSupersets}
\alias{fuzzyand}
\alias{fuzzyor}
\alias{getRow}
\alias{getSolution}
\alias{prettyString}
\alias{prettyTable}
\alias{print.aE}
\alias{print.DeMorgan}
\alias{print.fctr}
\alias{print.pic}
\alias{print.pof}
\alias{print.qca}
\alias{print.mqca}
\alias{print.tt}
\alias{rowDominance}
\alias{solveChart}
\alias{sortVector}
\alias{sortMatrix}
\alias{submodels.loop}
\alias{verify.ambig}
\alias{verify.data}
\alias{verify.dir.exp}
\alias{verify.inf.test}
\alias{verify.qca}
\alias{verify.mqca}
\alias{verify.tt}
\alias{writePrimeimp}
\alias{writeSolution}


\title{Internal Functions}

\description{
  These functions are internal; they are not intended to be called directly by the user.
}

\keyword{internal}